# backend/api/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "api.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/backend-dashboard/backend/export": {
            "post": {
                "summary": "/backend-dashboard/backend/export",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.backend.manager.service.v1.UrlResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.backend.manager.service.v1.ExpOptions"
                        }
                    }
                ],
                "tags": [
                    "backend.manager.service.v1.BackendManager"
                ]
            }
        },
        "/backend-dashboard/backend/models.delete": {
            "post": {
                "summary": "/backend-dashboard/backend/models.delete",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.backend.manager.service.v1.Model"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.backend.manager.service.v1.NameID"
                        }
                    }
                ],
                "tags": [
                    "backend.manager.service.v1.BackendManager"
                ]
            }
        },
        "/backend-dashboard/backend/models.insert": {
            "post": {
                "summary": "/backend-dashboard/backend/models.insert",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.backend.manager.service.v1.Model"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.backend.manager.service.v1.Model"
                        }
                    }
                ],
                "tags": [
                    "backend.manager.service.v1.BackendManager"
                ]
            }
        },
        "/backend-dashboard/backend/models.selectAll": {
            "post": {
                "summary": "/backend-dashboard/backend/models.selectAll",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.backend.manager.service.v1.ModelArray"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.backend.manager.service.v1.Empty"
                        }
                    }
                ],
                "tags": [
                    "backend.manager.service.v1.BackendManager"
                ]
            }
        },
        "/backend-dashboard/backend/models.selectByName": {
            "post": {
                "summary": "/backend-dashboard/backend/models.selectByName",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.backend.manager.service.v1.Model"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.backend.manager.service.v1.NameID"
                        }
                    }
                ],
                "tags": [
                    "backend.manager.service.v1.BackendManager"
                ]
            }
        },
        "/backend-dashboard/backend/models.update": {
            "post": {
                "summary": "/backend-dashboard/backend/models.update",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.backend.manager.service.v1.Empty"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.backend.manager.service.v1.Model"
                        }
                    }
                ],
                "tags": [
                    "backend.manager.service.v1.BackendManager"
                ]
            }
        }
    },
    "definitions": {
        ".backend.manager.service.v1.Empty": {
            "type": "object",
            "properties": {}
        },
        ".backend.manager.service.v1.ExpOptions": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "routePrefix": {
                    "type": "string"
                },
                "isMicoServ": {
                    "type": "boolean"
                }
            }
        },
        ".backend.manager.service.v1.Model": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "props": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/.backend.manager.service.v1.Prop"
                    }
                },
                "methods": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                }
            }
        },
        ".backend.manager.service.v1.ModelArray": {
            "type": "object",
            "properties": {
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/.backend.manager.service.v1.Model"
                    }
                }
            }
        },
        ".backend.manager.service.v1.NameID": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        ".backend.manager.service.v1.Prop": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        ".backend.manager.service.v1.UrlResp": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    }
}